# Use Memory<T> / Span<T>