--- a/Sources/Towel/Towel.csproj
+++ b/Sources/Towel/Towel.csproj
@@ -79,4 +79,8 @@ Generic mathematics, symbolic mathematics, generic path finding (graph search), 
     </Compile>
   </ItemGroup>
 
+  <ItemGroup>
+    <PackageReference Include="System.Memory" Version="4.5.3" />
+  </ItemGroup>
+
 </Project>