#### Description
Update all Array usages to use Memory<T> / Span<T>

#### Purpose
Better compatability & Performance

#### Issues
#48 

#### Mentions
@ZacharyPatten