@HurricanKai 2020-03-03
@ZacharyPatten how do you think `IArray<T>`/`Array<T>` should be handled.
I tbh don't understand the purpose of those anyways, should I leave the API as is, and still replace them with `Memory<T>` under the hood, rename them, or just leave them be?