@HurricanKai 2020-03-03
@ZacharyPatten looking at some more of the advanced code, these abstractions over built-in features like `GetIndex`, `SetIndex`, `Array` etc. seem pointless to me, and are hurting performance a lot.