@ZacharyPatten 2020-03-03
`Span<T>` and `Memory<T>` are nice but they do not abstract things enough. Having wrappers around getters and setters allow you to abstract the data structure being used in an algorithm. For example, the sorting algorithms work on arrays and lists, but they can also work on dictionary or any other `int` indexed data structure. It will even work on data structures made in other people's libraries.

The runtime delegate versions (`GetIndex<T` and `SetIndex<T>`) are slow for sure, but they are only syntax sugar for people who want to use the delegate/lambda style syntax. The more optimized approach is to make a `struct` rather than use a delegate. Then there is no delegate at runtime and it will potentially be inlined by the JIT.

I elaborate on this topic here: https://github.com/dotnet/csharplang/issues/2904

Also, the benchmarking code for the sorting algorithms is another example of how structs > delegates/lambdas for `static` function references: https://github.com/ZacharyPatten/Towel/blob/master/.github/Pages/Benchmarks.md