@ZacharyPatten 2020-03-03
Also, I never actually thought about using `Memory<T>` underneath the data structures. I was referring to other areas of the code when I commented that I want to use them.

I'm not opposed to using `Memory<T>` underneath the data structures if it works, but now I'm curious about stuff like serialization. I don't see the `Serializable` attribute on `Memory<T>`, and I would assume (knowing what Memory is) that it is not serializable.

That isn't necessarily a deal breaker, it just means custom serializers are required. The data structures with delegates underneath them already require custom serialization practices anyway.