@HurricanKai 2020-03-03
> Span<T> and Memory<T> are nice but they do not abstract things enough
I don't understand what you mean?
Only downside is that Span/Memory are linear only. but I don't see a use case where that is a problem.