@HurricanKai 2020-03-03
Also, when I say these add overhead, I mean a lot of it
As an example, bubble sort.
Before <https://github.com/ZacharyPatten/Towel/blob/444eea7d0da4a8a78410cd48b506a2a65ec7b23f/Sources/Towel/Sort.cs#L82>
After:
```cs
public static void New<T, TCompare>(Span<T> span, TCompare compare)
    where TCompare : IComparer<T>
{
    for (int i = 0; i < span.Length; i++)
    {
        for (int j = 0; j < span.Length - 1; j++)
        {
            if (compare.Compare(span[j], span[j + 1]) is 1)
            {
                var temp = span[j + 1];
                span[j + 1] = span[j];
                span[j] = temp;
            }
        }
    }
}
```

Benchmark:
```ini
BenchmarkDotNet=v0.12.0, OS=Windows 10.0.19569
AMD Ryzen 9 3900X, 1 CPU, 24 logical and 12 physical cores
.NET Core SDK=3.1.101
  [Host]     : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT
  Job-KFRPCW : .NET Framework 4.8 (4.8.4084.0), X64 RyuJIT
  Job-IXSZEK : .NET CoreRT 1.0.28703.02 @BuiltBy: dlab14-DDVSOWINAGE101 @Branch: master @Commit: 65f1d8fb61c8d4d426562f8861a89bfd24b173b4, X64 AOT

InvocationCount=1  UnrollFactor=1
```

|  Method |    Runtime | Size |             Mean |           Error |          StdDev |           Median |
|-------- |----------- |----- |-----------------:|----------------:|----------------:|-----------------:|
| Current |   .NET 4.8 |    2 |         454.2 ns |        31.77 ns |        91.67 ns |         500.0 ns |
|     New |   .NET 4.8 |    2 |         173.4 ns |        24.20 ns |        69.06 ns |         200.0 ns |
| Current | CoreRt 3.1 |    2 |         232.3 ns |        31.60 ns |        91.18 ns |         200.0 ns |
|     New | CoreRt 3.1 |    2 |         207.3 ns |        37.72 ns |       108.82 ns |         200.0 ns |
| Current |   .NET 4.8 |    8 |       1,607.4 ns |        40.11 ns |       115.08 ns |       1,600.0 ns |
|     New |   .NET 4.8 |    8 |         565.6 ns |        25.19 ns |        71.47 ns |         600.0 ns |
| Current | CoreRt 3.1 |    8 |         740.0 ns |        26.78 ns |        74.65 ns |         700.0 ns |
|     New | CoreRt 3.1 |    8 |         409.5 ns |        42.86 ns |       122.97 ns |         400.0 ns |
| Current |   .NET 4.8 |   16 |       6,247.4 ns |       124.98 ns |       138.92 ns |       6,300.0 ns |
|     New |   .NET 4.8 |   16 |       1,664.6 ns |        37.65 ns |        99.84 ns |       1,700.0 ns |
| Current | CoreRt 3.1 |   16 |       2,216.7 ns |        45.95 ns |        90.70 ns |       2,200.0 ns |
|     New | CoreRt 3.1 |   16 |       1,152.6 ns |        51.56 ns |       147.92 ns |       1,100.0 ns |
| Current |   .NET 4.8 |   32 |      18,814.3 ns |       145.79 ns |       129.24 ns |      18,800.0 ns |
|     New |   .NET 4.8 |   32 |       5,725.9 ns |       112.95 ns |       158.34 ns |       5,700.0 ns |
| Current | CoreRt 3.1 |   32 |       8,261.5 ns |       104.15 ns |        86.97 ns |       8,300.0 ns |
|     New | CoreRt 3.1 |   32 |       3,626.7 ns |        67.80 ns |       101.48 ns |       3,600.0 ns |
| Current |   .NET 4.8 |   64 |      79,161.5 ns |       562.84 ns |       470.00 ns |      79,100.0 ns |
|     New |   .NET 4.8 |   64 |      21,280.0 ns |       406.48 ns |       380.23 ns |      21,300.0 ns |
| Current | CoreRt 3.1 |   64 |      29,207.7 ns |       438.54 ns |       366.20 ns |      29,100.0 ns |
|     New | CoreRt 3.1 |   64 |      13,020.0 ns |       255.10 ns |       381.83 ns |      12,900.0 ns |
| Current |   .NET 4.8 |  128 |     326,438.5 ns |     2,859.31 ns |     2,387.66 ns |     326,700.0 ns |
|     New |   .NET 4.8 |  128 |      82,456.7 ns |     1,618.92 ns |     2,423.12 ns |      83,300.0 ns |
| Current | CoreRt 3.1 |  128 |     115,719.2 ns |       556.10 ns |       464.37 ns |     115,550.0 ns |
|     New | CoreRt 3.1 |  128 |      49,707.7 ns |       851.66 ns |       711.17 ns |      49,600.0 ns |
| Current |   .NET 4.8 |  256 |   1,260,860.0 ns |    13,735.23 ns |    12,847.95 ns |   1,256,100.0 ns |
|     New |   .NET 4.8 |  256 |     337,318.2 ns |     6,707.29 ns |    10,638.48 ns |     337,900.0 ns |
| Current | CoreRt 3.1 |  256 |     465,521.4 ns |     4,819.23 ns |     4,272.12 ns |     465,650.0 ns |
|     New | CoreRt 3.1 |  256 |     189,084.6 ns |       736.34 ns |       614.88 ns |     189,100.0 ns |
| Current |   .NET 4.8 |  512 |   5,292,871.4 ns |    43,246.91 ns |    38,337.26 ns |   5,294,200.0 ns |
|     New |   .NET 4.8 |  512 |   1,351,013.3 ns |    24,770.19 ns |    23,170.05 ns |   1,355,600.0 ns |
| Current | CoreRt 3.1 |  512 |   1,923,033.3 ns |    38,317.74 ns |    45,614.54 ns |   1,907,900.0 ns |
|     New | CoreRt 3.1 |  512 |     768,300.0 ns |     8,805.14 ns |     8,236.33 ns |     770,000.0 ns |
| Current |   .NET 4.8 | 1024 |  20,712,314.3 ns |   144,666.75 ns |   128,243.30 ns |  20,714,500.0 ns |
|     New |   .NET 4.8 | 1024 |   5,203,200.0 ns |   102,944.55 ns |   188,239.94 ns |   5,294,150.0 ns |
| Current | CoreRt 3.1 | 1024 |   7,803,133.3 ns |    98,370.99 ns |    92,016.28 ns |   7,785,400.0 ns |
|     New | CoreRt 3.1 | 1024 |   3,101,288.9 ns |    61,777.47 ns |    66,101.20 ns |   3,080,750.0 ns |
| Current |   .NET 4.8 | 2048 |  81,628,326.7 ns |   293,235.54 ns |   274,292.70 ns |  81,650,700.0 ns |
|     New |   .NET 4.8 | 2048 |  20,049,175.0 ns |   118,740.53 ns |    92,704.79 ns |  20,045,750.0 ns |
| Current | CoreRt 3.1 | 2048 |  30,131,207.7 ns |   125,104.01 ns |   104,467.51 ns |  30,115,900.0 ns |
|     New | CoreRt 3.1 | 2048 |  12,291,060.0 ns |   165,825.61 ns |   155,113.38 ns |  12,307,600.0 ns |
| Current |   .NET 4.8 | 4096 | 325,866,793.3 ns |   732,597.77 ns |   685,272.41 ns | 325,993,300.0 ns |
|     New |   .NET 4.8 | 4096 |  82,090,373.3 ns | 1,930,869.79 ns | 1,806,136.80 ns |  81,160,600.0 ns |
| Current | CoreRt 3.1 | 4096 | 121,594,757.1 ns |   246,554.96 ns |   218,564.55 ns | 121,579,500.0 ns |
|     New | CoreRt 3.1 | 4096 |  50,676,080.0 ns |   191,101.54 ns |   178,756.50 ns |  50,638,200.0 ns |